Changed energy calculation
can get new energy easily if only a small change is made to the state; in this
case if only one bit is flipped. so getenergy only called once, in main
I _think_ this gave roughly 10^3 speedup
it's a lot, but need to simulate lot more spins because:

But still some weird things going on, probability of flipping is always too
high, think Ediff not being calculated properly
^^This was because we actually have 4*cellsize^3 atoms and entries in intmat,
not 16*cellsize^3, so N was off

do i even need to keep track of energy? no